restructure root folder
