+++ Digraphy.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Digraphy", "Digraphy\Digraphy.csproj", "{2EB4BB0B-E4BD-4143-9A5B-9BDA6333B7E0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2EB4BB0B-E4BD-4143-9A5B-9BDA6333B7E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2EB4BB0B-E4BD-4143-9A5B-9BDA6333B7E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2EB4BB0B-E4BD-4143-9A5B-9BDA6333B7E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2EB4BB0B-E4BD-4143-9A5B-9BDA6333B7E0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
