--- Digraphy/.idea/.idea.Digraphy/.idea/indexLayout.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="UserContentModel">
-    <attachedFolders>
-      <Path>../../../Digraphy</Path>
-    </attachedFolders>
-    <explicitIncludes />
-    <explicitExcludes />
-  </component>
-</project>
