+++ DigraphyApi/Program.cs
@@ -1,0 +1,40 @@
+﻿using Digraphy.Data;
+using Digraphy.Interfaces;
+using Digraphy.Mapper;
+using Digraphy.Repository;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+var builder = WebApplication.CreateBuilder(args);
+
+builder.Services.AddControllers();
+builder.Services.AddAutoMapper(typeof(MappingProfiles));
+builder.Services.AddScoped<ITodoRepository, TodoRepository>();
+builder.Services.AddDbContextPool<AppDbContext>(opt =>
+    opt.UseNpgsql(builder.Configuration.GetConnectionString("DigraphyDatabase")));
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+var app = builder.Build();
+
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI(options =>
+    {
+        // Update the SwaggerUI configuration to specify the base URL
+        options.SwaggerEndpoint("/swagger/v1/swagger.json", "DigraphyApi");
+        options.RoutePrefix = string.Empty;
+    });
+}
+
+// app.UseHttpsRedirection();
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();
